Refactor: Separate out SimpleHashClient from AssetDiscoveryTask
and add SimpleHashClient as a member of AssetDiscoveryTask as well as
BraveWalletService.  Previously only AssetDiscoveryTask needed to
interface with SimpleHash, however the new
BraveWalletService::GetSpamNFTs API will also use interface with
SimpleHash, and as a result I've separated out this logic.

Also, scaffold an empty implementation for
BraveWalletService::GetSpamNFTs.
