+++ components/brave_wallet/browser/asset_discovery_manager.cc
@@ -50,6 +50,7 @@ AssetDiscoveryManager::AssetDiscoveryManager(
     BraveWalletService* wallet_service,
     JsonRpcService* json_rpc_service,
     KeyringService* keyring_service,
+    SimpleHashClient* simple_hash_client,
     PrefService* prefs)
     : api_request_helper_(std::make_unique<APIRequestHelper>(
           GetAssetDiscoveryManagerNetworkTrafficAnnotationTag(),
@@ -57,6 +58,7 @@ AssetDiscoveryManager::AssetDiscoveryManager(
       wallet_service_(wallet_service),
       json_rpc_service_(json_rpc_service),
       keyring_service_(keyring_service),
+      simple_hash_client_(simple_hash_client),
       prefs_(prefs),
       weak_ptr_factory_(this) {
   keyring_service_->AddObserver(
@@ -157,7 +159,8 @@ void AssetDiscoveryManager::AddTask(
   auto non_fungible_supported_chains = GetNonFungibleSupportedChains();
 
   auto task = std::make_unique<AssetDiscoveryTask>(
-      api_request_helper_.get(), wallet_service_, json_rpc_service_, prefs_);
+      api_request_helper_.get(), simple_hash_client_, wallet_service_,
+      json_rpc_service_, prefs_);
   auto callback = base::BindOnce(&AssetDiscoveryManager::FinishTask,
                                  weak_ptr_factory_.GetWeakPtr());
   auto* task_ptr = task.get();
