+++ components/brave_wallet/browser/brave_wallet_service.h
@@ -18,6 +18,7 @@
 #include "brave/components/brave_wallet/browser/asset_discovery_manager.h"
 #include "brave/components/brave_wallet/browser/brave_wallet_p3a.h"
 #include "brave/components/brave_wallet/browser/brave_wallet_service_delegate.h"
+#include "brave/components/brave_wallet/browser/simple_hash_client.h"
 #include "brave/components/brave_wallet/common/brave_wallet.mojom.h"
 #include "components/keyed_service/core/keyed_service.h"
 #include "components/prefs/pref_change_registrar.h"
@@ -50,6 +51,7 @@ class BraveWalletService : public KeyedService,
                            public mojom::BraveWalletService,
                            public BraveWalletServiceDelegate::Observer {
  public:
+  using APIRequestHelper = api_request_helper::APIRequestHelper;
   using SignMessageRequestCallback =
       base::OnceCallback<void(bool,
                               mojom::ByteArrayStringUnionPtr,
@@ -233,6 +235,11 @@ class BraveWalletService : public KeyedService,
   void GetBalanceScannerSupportedChains(
       GetBalanceScannerSupportedChainsCallback callback) override;
 
+  void GetSpamNFTs(const std::string& wallet_address,
+                   const std::vector<std::string>& chain_ids,
+                   const absl::optional<std::string>& cursor,
+                   GetSpamNFTsCallback callback) override;
+
   // BraveWalletServiceDelegate::Observer:
   void OnActiveOriginChanged(const mojom::OriginInfoPtr& origin_info) override;
 
@@ -374,6 +381,7 @@ class BraveWalletService : public KeyedService,
   raw_ptr<TxService> tx_service_ = nullptr;
   raw_ptr<PrefService> profile_prefs_ = nullptr;
   BraveWalletP3A brave_wallet_p3a_;
+  std::unique_ptr<SimpleHashClient> simple_hash_client_;
   std::unique_ptr<AssetDiscoveryManager> asset_discovery_manager_;
   std::unique_ptr<EthAllowanceManager> eth_allowance_manager_;
   mojo::ReceiverSet<mojom::BraveWalletService> receivers_;
