+++ components/brave_wallet/browser/test/BUILD.gn
@@ -52,6 +52,7 @@ source_set("brave_wallet_unit_tests") {
     "//brave/components/brave_wallet/browser/permission_utils_unittest.cc",
     "//brave/components/brave_wallet/browser/rlp_decode_unittest.cc",
     "//brave/components/brave_wallet/browser/rlp_encode_unittest.cc",
+    "//brave/components/brave_wallet/browser/simple_hash_client_unittest.cc",
     "//brave/components/brave_wallet/browser/simulation_request_helper_unittest.cc",
     "//brave/components/brave_wallet/browser/simulation_response_parser_unittest.cc",
     "//brave/components/brave_wallet/browser/simulation_service_unittest.cc",
